Add redux, add <Landing /> screen
