+++ App.js
@@ -1,23 +1,19 @@
 import React from 'react';
-import { StyleSheet, Text, View } from 'react-native';
+import { connect, Provider } from 'react-redux';
+import { createStore } from 'redux';
+import rootReducer from 'src/reducers';
+import Landing from 'src/screens/Landing';
 
-export default class App extends React.Component {
+
+class Main extends React.Component {
   render() {
+    const store = createStore(rootReducer);
     return (
-      <View style={styles.container}>
-        <Text>Open up App.js to start working on your app!</Text>
-        <Text>Changes you make will automatically reload.</Text>
-        <Text>Shake your phone to open the developer menu.</Text>
-      </View>
+      <Provider store={store}>
+        <Landing />
+      </Provider>
     );
   }
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
+export default Main
