+++ package.json
@@ -3,12 +3,13 @@
   "version": "0.1.0",
   "private": true,
   "devDependencies": {
-    "react-native-scripts": "1.11.1",
     "jest-expo": "25.0.0",
+    "react-native-scripts": "1.11.1",
     "react-test-renderer": "16.2.0"
   },
   "main": "./node_modules/react-native-scripts/build/bin/crna-entry.js",
   "scripts": {
+    "format": "prettier --trailing-comma=es5 --single-quote --write \"src/**/*.js\"",
     "start": "react-native-scripts start",
     "eject": "react-native-scripts eject",
     "android": "react-native-scripts android",
@@ -20,7 +21,10 @@
   },
   "dependencies": {
     "expo": "^25.0.0",
+    "prettier": "^1.10.2",
     "react": "16.2.0",
-    "react-native": "0.52.0"
+    "react-native": "0.52.0",
+    "react-redux": "^5.0.6",
+    "redux": "^3.7.2"
   }
 }
