+++ src/package.json
@@ -1,0 +1,1 @@
+{"name": "src"}
