+++ src/screens/Landing/index.js
@@ -1,0 +1,28 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { StyleSheet, Text, View } from 'react-native';
+
+export default class Landing extends React.Component {
+  static propTypes = {};
+  static defaultProps = {};
+
+  render() {
+    return (
+      <View style={styles.container}>
+        <Text style={styles.text}>Hello Oscars 2018</Text>
+      </View>
+    );
+  }
+}
+
+const styles = StyleSheet.create({
+  container: {
+    alignItems: 'center',
+    flex: 1,
+    justifyContent: 'center',
+  },
+  text: {
+    color: 'black',
+    fontSize: 25,
+  },
+});
